lws_client_connect_via_info API is not multi-thread reentrant. Simply interlocking it will lead to live locks.
LibWebSockets assumes a single threaded execution. While we are connected to websocket we need to run ICE server info refresh which requires interlacing a few of the API calls with the callbacks.
This leaves us with two choices - either disconnect from the connected websocket when we do ICE refresh and reconnect it once that's done or to spin up two separate contexts. The first solution is more straight forward but at the expense of potentially losing a message (if it falls out of TTL) or having high latency when viewers are connecting. The dual context solution is bad because it doesn't scale down to small footprint devices. So we decided to use the same context to drive synchronous in-flight requests.
The interlocking is not a trivial thing to have as it leads to live-locks in case where the service thread is blocked on a socket and has the lock and there is no way to wake it up as the wake up API itself would require a lock. The problem is more exacerbated by the fact that we could have multiple entities in-flight - connection listener, ICE refresh and the reconnect thread and they need to be carefully coordinated and interlocked during the negative cases and teardown.

The solution here is to store the WSI when connecting in an interlocked fashion which uses a separate serialization lock and not the service lock to avoid live locking. While acquiring the lock, we try to ackquire it and if we fail, that means another thread is locked (for example the listener thread servicing a websocket while we are executing ICE refresh). As the holder can be blocked on the socket, we need to "wake it up" without having the lock held. Apparently, there are only TWO APIs that can do that. From https://libwebsockets.org/lws-api-doc-master/html/md_READMEs_README_8coding.html

<
Therefore the two apis mentioned above that may be used from another thread are

For LWS using the default poll() event loop, lws_callback_on_writable()
For LWS using libuv/libev/libevent event loop, lws_cancel_service()
>