Optimize startup latency from ~9s to ~80ms
The startup latency is measured with the following procedures:
  * Modified the JS example, https://awslabs.github.io/amazon-kinesis-video-streams-webrtc-sdk-js/examples/index.html,
    which basically listen to "loadeddata" event which basically it lets us
    know when the video starts flowing on the UI. When this event happens,
    it send a data channel message to the Webrtc C SDK
  * Modified the Webrtc C sample to measure the duration between connected
    to the time when it receives the data channel message from the UI

In my test, I got ~9 s delay. After the fix, the delay became ~80 ms.

The issue was that we started sending frames when the connection was still
not ready. Consequently, the first I-frame was lost, and the receiver needs
to wait until the next I-frame. So, for the delay itself, it'll be varied
depending on the configured keyframe interval. With our stock sample, I
calculated the delay to be ~8 s, which is matched to the observed delay.

The details about the sample video:
  * Keyframe interval is 250
  * Frame lost: first 16 frames
  * Frame per second: 30
  * Initial delay: (Keyframe interval - frame lost) / frame per second = 7.8 s