Update createRtpPacketFromBytes(), fix double free
createRtpPacketFromBytes() now only moves ownership of the rawPacket
argument to the PRtpPacket object if the call to
createRtpPacketFromBytes() is successful. This allows the caller to
free rawPacket in a CleanUp section without having to check if
a call to createRtpPacketFromBytes() was the source of failure.

This commit also fixes two potential double frees in
sendPacketToRtpReceiver().

This first potential double free is that the call to
createRtpPacketFromBytes() in sendPacketToRtpReceiver() succeeds, but
then the call to jitterBufferPush() fails. When this happens, then
SAFE_MEMFREE() frees pPayload. However, the subsequent call to
freeRtpPacket() then also calls free on pPayload since
createRtpPacketFromBytes() transferred ownership of pPayload to
pRtpPacket.

This commit fixes the first potential double free by setting pPayload
to NULL after ownership of pPayload is transferred to pRtpPacket.

The second potential double free is that the call to
createRtpPacketFromBytes() in sendPacketToRtpReceiver() fails. When
this happens, the call to SAFE_MEMFREE() in the CleanUp section
performs a double free because createRtpPacketFromBytes() already
freed pPayload.

This commit fixes the second potential double free by changing
createRtpPacketFromBytes() only to transfer ownership when
createRtpPacketFromBytes() is successfully called, i.e.,
createRtpPacketFromBytes() won't free the rawPacket argument when the
call fails.