Add network interface filter example in samples (#1176)
* Add network interface filter example in samples

* Try removing alpine update

* Update .travis.yml

It looks like the issue is docker isn't up to date because it has an older version of `libseccomp`, see here:
 https://wiki.alpinelinux.org/wiki/Release_Notes_for_Alpine_3.13.0#time64_requirements

I found instructions for how to update docker in travis ci using addon.  I found that to user docker we're supposed to be specifying it in a "services:" sections which we weren't doing but it was working somehow, regardless I've added it.  Not sure if it'll work but here are the directions to update docker on travis ci:  https://docs.travis-ci.com/user/docker/#installing-a-newer-docker-version

* Test travis ci

* Debug windows build by commenting callback

* Disabling all builds except Windows

* Adding one argument...incremental

* Add argument 2...incremental

* fix name of identifier

* Test with old version of libwebsockets to ensure Windows is tested

* Update with usrsctp and static build changes

* Network interface filter function sample

Co-authored-by: Hassan Sahibzada <hsahibza@amazon.com>