Fix compilation of OpenSSL 1.1.1t on Raspberry Pi 4B with 64-bit kernel, 32-bit userland.
$ uname -a
Linux raspberrypi 6.1.21-v8+ #1642 SMP PREEMPT Mon Apr  3 17:24:16 BST 2023 aarch64 GNU/Linux
$ file $(which file)
/usr/bin/file: ELF 32-bit LSB executable, ARM, EABI5 version 1 (SYSV), dynamically linked, interpreter /lib/ld-linux-armhf.so.3, BuildID[sha1]=46998c6200b7bcc2d6b1539b82d4044903017bb4, for GNU/Linux 3.2.0, stripped

The OpenSSL 1.1.1t Configure script selects aarch64 instead of linux-armv4 as
the machine configuration in this case, this results in a build failure when
the 32-bit assembler is provided with 64-bit opcodes.

If the OpenSSL architecture is not explicitly set, then if the
architecture is aarch64, a test is done to see if userland is 32 bit or
64 bit. If it is 32 bit, then OpenSSL is configured with the linux-armv4
architecture instead. This results in assembly language that works.